+++ anyblok/bloks/anyblok_core/core/sqlbase.py
@@ -557,15 +557,3 @@ class SqlBase(SqlMixin):
             cls.registry.flush()
 
         return instances
-
-    @classmethod
-    def precommit_hook(cls, method, *args, **kwargs):
-        """ Same in the registry a hook to call just before the commit
-
-        .. warning:: Only one instance of the hook is called before the commit
-
-        :param method: the method to call on this model
-        :param put_at_the_end_if_exist: If ``True`` the hook is move at the end
-        """
-        cls.registry.precommit_hook(
-            cls.__registry_name__, method, *args, **kwargs)