+++ doc/CHANGES.rst
@@ -12,6 +12,13 @@
 CHANGELOG
 =========
 
+Futur
+-----
+
+* [ADD] log debug for commit / rollback
+* [REF] precommit_hook, can also be on no SQL Model
+* [ADD] postcommit_hook
+
 0.10.1 (2017-10-14)
 -------------------